Update 20230522 - Improvement to Service Archiving Logic
Improved the logic for identifying an essential service (from commit 2771ef2). This is essentially a temporary stop gap which has lots of flaws and could lead to a doomsday scenario where the variation cannot be deleted and will be in the system until the heat of the universe dies off. Improvements on transaction is recommended, such as completely making the transactions full on dynamic to allow cleaner database which utilizes a single table instead of multiple.

[CHANGELOG]
🟢 Added a query that identifies the ID of boarding service category
🟡 Moved the boarding service identifier logic from the outside, to the initialization part of the array; utilizing the ID from the new logic query to fetch the oldest entry of boarding and thus, preventing it from being archived.